Change design of inputs in the settings tab
