+++ src/pages/Profile/_Profile.scss
@@ -1,7 +1,7 @@
 @import "./Sidebar/Sidebar";
 @import "./ProfileNav";
 @import "./subpages/Settings";
-@import "./subpages/Settings/Contact.scss";
+@import "./subpages/Settings/Social.scss";
 @import "./subpages/Settings/Security.scss";
 @import "./subpages/Settings/Basics.scss";
 @import "./subpages/Settings/Traveler.scss";
