+++ src/pages/Profile/subpages/Settings/Basics.scss
@@ -2,7 +2,7 @@
   display: flex;
   flex-wrap: wrap;
   margin: 15px;
-
+  display: block;
   .input-container {
     margin: 10px 0 10px 0;
     width: 400px;
@@ -16,7 +16,7 @@
     }
 
     .error {
-      color: red;
+      color: #b10606;
       font-size: 0.9em;
     }
 
@@ -30,5 +30,11 @@
     [type="date"]::-webkit-calendar-picker-indicator {
       opacity: 0;
     }
+  }
+
+  .button {
+    margin: 10px;
+    display: block;
+    width: fit-content;
   }
 }
