+++ src/pages/Profile/subpages/Settings/Security.scss
@@ -20,28 +20,6 @@
       margin-bottom: 8px;
       min-width: 300px;
     }
-    .security-password-input {
-      color: rgb(210, 224, 233);
-      font-family: "Segoe UI", sans-serif;
-      font-size: 16px;
-      background: #4f5867;
-      border: none;
-      border-radius: 16px;
-      padding-left: 12px;
-      box-shadow: 0 1px 1px 2px rgba(0, 0, 0, 0.15);
-      height: 32px;
-      width: 200px;
-      &:focus {
-        border: 2px solid #ccc; 
-        outline: none;
-      }
-      @media (min-width: 900px) {
-        width: 300px;
-      }
-      @media (min-width: 1000px) {
-        width: 400px;
-      }
-    }
 
     span.validate {
       margin: 4px;
@@ -50,30 +28,8 @@
       margin-bottom: -8px;
     }
 
-    input.submit-button {
+    input.button {
       margin-top: 20px;
-      cursor: pointer;
-      
-    }
-  }
-
-  .security-delete-button {
-    margin: 10px 0 0 0;
-    cursor: pointer;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 200px;
-    min-height: 40px;
-    border-radius: 4px;
-    font-weight: 550;
-    font-size: 18px;
-    background: #ecd7db;
-    color: #bb6f7e;
-    box-shadow: 0 1px 1px 1px rgba(0, 0, 0, 0.15);
-    @media (min-width: 1000px) {
-      width: 300px;
-      font-size: 24px;
     }
   }
 }
